backend/requirements.txt
flask
flask-cors
tensorflow
pandas
numpy
python-dotenv
requests
